@qRoC 2024-09-02
> When the Reuseable<T> is dropped in foreground thread, it should not be reused because of the synchronization overhead.

Since we now have vectors with capacity in the global pool, we start to choose between a small overhead on the mutex or an overhead on allocations/reallocations.

I see #224 and foyer-rs/foyer#571, the main problem that was - when the thread does not use attributes, we spend a lot of time on synchronization. But now such threads will work as with the `REUSABLE` flag, because having zero capacity they will not synchronize.

> thread pool can be the same as global pool: only with capacity

Yes, but attributes must be  with`Option`, and main problem - current implementation does not require `TrustedLen` in API.

> Instead of ReusableVec::stub(), I recommend to create a Vec::new() which is zero-overhead and wrap it in Reusable.

`stub` do that.