@andylokandy 2024-09-04
> we start to choose between a small overhead on the mutex or an overhead on allocations/reallocations

I don't have a number on the overhead of the mutex when it's in heavy conflict situation, and it may cause long tail on worst case. On the contrary, dropping the vector without reuse is rare, and reallocating large vector when collecting large amount of spans is acceptable and will not cause long tail.

> I see #224 and https://github.com/foyer-rs/foyer/pull/571, the main problem that was - when the thread does not use attributes, we spend a lot of time on synchronization. But now such threads will work as with the REUSABLE flag, because having zero capacity they will not synchronize.

It's not totally correct. The real case is that, when `LocalParentGuard` drops, it drops the `Reusable` and causes synchronization. In the new version, `LocalParentGuard` always submit the `Reusable` to global collector so that the vectors are reused in the global collector thread -- result in no synchronization overhead.

In short, with `REUSABLE` flag, the sync should never happen. Reuse should only take place in the global collector thread.

Did this PR resolve this problem in the other way?